Allow disabling of docs target
When we do not have doxygen installed and do not want to build docs, this gives out unnecessary warning in cmake:
```
WARNING: Doxygen not found, so disabling quickcpplib_docs target quickcpplib/cmakelib/QuickCppLibMakeDoxygen.cmake:4 (indented_message) quickcpplib/CMakeLists.txt:42 (include)
```

This allows disabling the creation of quickcpplib_docs target if not needed.

The original behavior is preserved, the docs target is created and built by default.