build: Remove dependence on include paths for configs
Due to the requirement that the sources can be copy pasted into another
project without _any_ build system support (including properly setup
include paths), we split the configuration inputs into two separate
systems. For in-tree usage we supply the configuration options per
commandline, while for the installed source we write them to a config
header file. The source tree contains an empty placeholder in order to
avoid the need for conditional inclusion logic.